% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_get_methods.R
\name{all_get_methods}
\alias{all_get_methods}
\title{All Public GET Methods}
\usage{
all_get_methods()
}
\value{
a \code{tbl_df} of API \code{GET} methods
}
\description{
Retrieve a list of all the available public \code{GET} methods
}
