% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-files.R
\name{create.MAF}
\alias{create.MAF}
\title{Create a dataframe representing a MAF file}
\usage{
create.MAF(assaytype = "LCMS", nrow = NULL, abundances = NULL)
}
\arguments{
\item{assaytype}{Currently unused, but would be used for different MAF files for NMR and MS}

\item{nrow}{Create the MAF with nrow empty rows.}

\item{abundances}{is a data frame or matrix with the metabolite abundances,
intensities or concentrations.}
}
\description{
MAF files combine the metabolite abundance matrix,
the spectral data like m/z and retention time,
and the metabolite names, structures and database identifiers.
}
\details{
If an abundance matrix is given, the the generated MAF will have
as many (empty) spectral and identification metadata rows as abundance rows,
joined to the provided abundance matrix.

nrow and abundances should be mutually exclusive, but that is not yet tested.
}
\examples{
maf <- create.MAF(nrow=17)
}
