% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_study_file.R
\name{download_study_file}
\alias{download_study_file}
\title{Download Study File Contents}
\usage{
download_study_file(study_id, filename)
}
\arguments{
\item{study_id}{a character string of a valid MTBLS study id}

\item{filename}{a character of the full filename and extension to download}
}
\value{
a \code{tibble} of file contents
}
\description{
Download the contents of a specified file from a public study
}
