% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isa_sample.R
\name{get_isa_sample}
\alias{get_isa_sample}
\title{Get ISA Sample File}
\usage{
get_isa_sample(study_id, filename, outpath, overwrite = FALSE)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{outpath}{A character string of a filepath where the ISA sample `.txt` file will be saved to}

\item{overwrite}{Default FALSE. Set to TRUE to overwrite an existing file.}

\item{the}{study filename fo the ISA sample file}
}
\description{
Get ISA Sample File
}
