% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_private_studies.R
\name{get_private_studies}
\alias{get_private_studies}
\title{Get Private Studies}
\usage{
get_private_studies()
}
\value{
a \code{tibble} of your private Study IDs
}
\description{
Retrieve all private studies which belong to your user account
}
\examples{
\dontrun{
get_private_studies()
}
}
