% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_file.R
\name{get_raw_file}
\alias{get_raw_file}
\title{Get raw file}
\usage{
get_raw_file(study_id, filename, outpath, overwrite = FALSE)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{filename}{A character string of a study file to read as a table}

\item{outpath}{A character string of a path to save the raw file}

\item{overwrite}{Default FALSE. Set to TRUE to overwrite an existing file.}
}
\description{
Get raw file
}
