% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_studies.R
\name{get_studies}
\alias{get_studies}
\title{Get Studies}
\usage{
get_studies()
}
\value{
a \code{tibble} of Study IDs
}
\description{
Retrieve a list of all publically available Metabolights studies
}
\examples{
get_studies()
}
