% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_assay_file.R
\name{get_study_assay_file}
\alias{get_study_assay_file}
\title{Get Study Assay File}
\usage{
get_study_assay_file(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a character string of the study assay filename
}
\description{
Retrieve the filename details for available assays of a public study
}
\examples{
get_study_assay_file('MTBLS375')
}
