% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_assay_list.R
\name{get_study_assay_list}
\alias{get_study_assay_list}
\title{Get Study Description}
\usage{
get_study_assay_list(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a list off the study assays and associated details
}
\description{
Retrieve the abstract based description of a public study
}
\examples{
get_study_assay_list('MTBLS375')
}
