% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_audit.R
\name{get_study_audit}
\alias{get_study_audit}
\title{Get Study Audit}
\usage{
get_study_audit(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{list} of available audit folders
}
\description{
Retrieve a list of available audit folders for a study
}
\examples{
get_user_studies
}
