% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_contacts.R
\name{get_study_contacts}
\alias{get_study_contacts}
\title{Get Study Contacts}
\usage{
get_study_contacts(study_id)
}
\arguments{
\item{study_id}{a character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of study contacts
}
\description{
Retrieve the contact details for a specified public study
'
}
\examples{
\dontrun{
get_study_contacts('MTBLS375')
}
}
