% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_desc.R
\name{get_study_desc}
\alias{get_study_desc}
\title{Get Study Description}
\usage{
get_study_desc(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a character string of the study description
}
\description{
Retrieve the abstract based description of a public study
}
\examples{
get_study_desc('MTBLS375')
}
