% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_descriptors.R
\name{get_study_descriptors}
\alias{get_study_descriptors}
\title{Get Study Descriptors}
\usage{
get_study_descriptors(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of study descriptors
}
\description{
Retrieve the study descriptors and annotation values for a publically available study
}
\examples{
\dontrun{
get_study_descriptors('MTBLS375')
}
}
