% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_factors.R
\name{get_study_factors}
\alias{get_study_factors}
\title{Get Study Factors}
\usage{
get_study_factors(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a `tibble` of study factors
}
\description{
Get Study Factors
}
