% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_files.R
\name{get_study_files}
\alias{get_study_files}
\title{Get Study Files}
\usage{
get_study_files(study_id, raw_data = FALSE)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{raw_data}{logical; if `TRUE` then raw data file info is also returned. (Default is `FALSE`)}
}
\value{
a `tibble` of file info
}
\description{
Get Study Files
}
