% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_meta.R
\name{get_study_meta}
\alias{get_study_meta}
\title{Get Study Meta}
\usage{
get_study_meta(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a `tibble` of study meta data
}
\description{
Get Study Meta
}
