% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_org.R
\name{get_study_org}
\alias{get_study_org}
\title{Get Study Organisms}
\usage{
get_study_org(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of study organism data
}
\description{
Retrieve the organisms used within a study
}
\examples{
\dontrun{
get_study_org('MTBLS375')
}
}
