% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_protocols.R
\name{get_study_protocols}
\alias{get_study_protocols}
\title{Get Study Protocols}
\usage{
get_study_protocols(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of study protocols
}
\description{
Retrieve the protocol information that has been deposited for a public study
}
\examples{
get_study_protocols('MTBLS375')
}
