% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_samples.R
\name{get_study_samples}
\alias{get_study_samples}
\title{Get Study Samples}
\usage{
get_study_samples(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of filename sample name and reliability score. 1.0 indicates a perfect match
}
\description{
Get a list of all sample names mapped to files within the study
}
\examples{
\dontrun{
get_study_samples('MTBLS375')
}
}
