% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_tech.R
\name{get_study_tech}
\alias{get_study_tech}
\title{Get Study Technology}
\usage{
get_study_tech()
}
\value{
a \code{tibble} of study id and technology
}
\description{
Retrieve a \code{tibble} of all Study IDs and the analytical technology used in the study
}
\examples{
get_study_tech()
}
