% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_title.R
\name{get_study_title}
\alias{get_study_title}
\title{Get Study Title}
\usage{
get_study_title(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a character string of the study title
}
\description{
Retrieve the full title of the study
}
\examples{
\dontrun{
get_study_title('MTBLS375')
}
}
