% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tsv.R
\name{get_tsv}
\alias{get_tsv}
\title{Get tsv file from study as a tibble}
\usage{
get_tsv(study_id, tsv_file)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{tsv_file}{A character string of the filename to get}
}
\value{
a `tibble` of tsv data
}
\description{
Get tsv file from study as a tibble
}
