% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_studies.R
\name{get_user_studies}
\alias{get_user_studies}
\title{Get Studies for User}
\usage{
get_user_studies()
}
\value{
a \code{list} of user owned studies
}
\description{
Retrieve a list of studies for a user
}
\examples{
get_user_studies
}
