% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_webservice.R
\name{get_webservice}
\alias{get_webservice}
\title{Get Web-service details}
\usage{
get_webservice()
}
\value{
a list of the MTBLS webservice information
}
\description{
Retrieve details about the Metabolights RESTful WebService
}
\examples{
get_webservice()
}
