% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_token.R
\name{mtbls_key}
\alias{mtbls_key}
\title{Set API Token}
\usage{
mtbls_key(API_KEY = NULL)
}
\arguments{
\item{API_KEY}{a character string of your Metabolights API Token (Default is `NULL`)}
}
\description{
Set your Metabolights API Token as global option. If `MTBLS_API_KEY` is set in `.Renviron` then this variable will be passed directly
to the `MTBLS_API_KEY` option. If no `MTBLS_API_KEY` is set in `.Renviron`, then the API Token must be passed as an input in the `set_api_token`
function
}
\examples{
\dontrun{

# If API Token is set in `.Renviron`

mtbls_key()

# If API Token is not set in `.Renviron`

mtbls_key('XXXX-0000-XXXX-0000')
}
}
