% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{protocols_annotation}
\alias{protocols_annotation}
\title{Parse Protocol Annotation Values}
\usage{
protocols_annotation(x)
}
\arguments{
\item{x}{the \code{httr} parsed output from \code{get_study_protocols}}
}
\value{
a character string of protocol annotations
}
\description{
Parse Protocol Annotation Values
}
\keyword{internal}
