% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metabolite.R
\name{search_metabolite}
\alias{search_metabolite}
\title{Search Metabolite}
\usage{
search_metabolite(search_name)
}
\arguments{
\item{search_name}{A character string of a chemical name to search for}
}
\description{
Search Metabolite
}
\examples{
search_metabolite('proline')
}
