% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf-files.R
\name{write.MAF}
\alias{write.MAF}
\title{Write a dataframe to a MAF file}
\usage{
write.MAF(maf, file = "")
}
\arguments{
\item{maf}{MAF data frame to write}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}
}
\description{
Write a suitably formatted data frame to a MAF file.
}
\examples{
maf <- create.MAF(nrow=17)
write.MAF(maf, file=tempfile())
}
