% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isa_investigation.R
\name{get_isa_investigation}
\alias{get_isa_investigation}
\title{Get ISA Investigation File}
\usage{
get_isa_investigation(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
the ISA-tab Investigation output
}
\description{
Retrieve the ISA Investigation for a specified Metabolights Study
}
\examples{
\dontrun{
ISA <- get_isa_investigation('MTBLS375')
writeLines(isa_investigation_parse, con = paste0('outpath/ISA.txt'))
}
}
