% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isa_investigation.R
\name{get_isa_investigation}
\alias{get_isa_investigation}
\title{Get ISA Investigation File}
\usage{
get_isa_investigation(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
the ISA-tab Investigation output
get_isa_investigation(study_id, outpath, overwrite = FALSE)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{outpath}{A character string of a filepath where the ISA Investigation `.txt` file will be saved to}

\item{overwrite}{Default FALSE. Set to TRUE to overwrite an existing file.}

