% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_files.R
\name{get_study_files}
\alias{get_study_files}
\title{Get Study Files}
\usage{
get_study_files(study_id, raw_data = FALSE)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}

\item{raw_data}{logical; if \code{TRUE} then raw data file info is also returned. (Default is \code{FALSE})}
}
\value{
a \code{tibble} of file info
}
\description{
Retrieve the names and type of all files that have been deposited with the public study. If \code{raw_data} is TRUE then details of raw data files(ie, \code{.mzML}) are also returned.
}
\examples{
\dontrun{
get_study_files('MTBLS375', raw_data = FALSE)
}
}
