% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_pubs.R
\name{get_study_pubs}
\alias{get_study_pubs}
\title{Get Study Publications}
\usage{
get_study_pubs(study_id)
}
\arguments{
\item{study_id}{A character string of a valid MTBLS study id}
}
\value{
a \code{tibble} of study publications
}
\description{
Retrieve details on any publications that have been associated with the study
}
\examples{
get_study_pubs('MTBLS375')
}
