% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbls_key.R
\name{mtbls_key}
\alias{mtbls_key}
\title{Set API Token}
\usage{
mtbls_key(API_KEY = NULL)
}
\arguments{
\item{API_KEY}{a character string of your Metabolights API Token (Default is \code{NULL})}
}
\description{
Set your Metabolights API Token as global option. If \code{MTBLS_API_KEY} is set in \code{.Renviron} then this variable will be passed directly
to the \code{MTBLS_API_KEY} option. If no \code{MTBLS_API_KEY} is set in \code{.Renviron}, then the API Token must be passed as an input in the \code{set_api_token}
function
}
\examples{
\dontrun{

# If API Token is set in `.Renviron`

mtbls_key()

# If API Token is not set in `.Renviron`

mtbls_key('XXXX-0000-XXXX-0000')
}
}
